refactor: todo_list.pyをアプリのエントリーポイントに変更する #14
